split layouts into shared/ui
