+++ packages/fields-document/src/DocumentEditor/layouts.tsx
@@ -1,8 +1,9 @@
 /** @jsxRuntime classic */
 /** @jsx jsx */
+'use client'
 
 import { createContext, useContext, useMemo } from 'react'
-import { Editor, Element, Node, Transforms, Range, Point } from 'slate'
+import { Transforms } from 'slate'
 import {
   type RenderElementProps,
   ReactEditor,
@@ -14,29 +15,26 @@ import {
 import { jsx, useTheme } from '@keystone-ui/core'
 import { Tooltip } from '@keystone-ui/tooltip'
 import { Trash2Icon } from '@keystone-ui/icons/icons/Trash2Icon'
-
 import { ColumnsIcon } from '@keystone-ui/icons/icons/ColumnsIcon'
 import { useControlledPopover } from '@keystone-ui/popover'
+
 import { type DocumentFeatures } from '../views-shared'
 import { InlineDialog, ToolbarButton, ToolbarGroup, ToolbarSeparator } from './primitives'
-import { paragraphElement } from './paragraphs'
-import {
-  insertNodesButReplaceIfSelectionIsAtEmptyParagraphOrHeading,
-  isElementActive,
-  moveChildren,
-} from './utils'
+import { isElementActive, } from './utils'
 import { useToolbarState } from './toolbar-state'
+
+import { insertLayout } from './layouts-shared'
 
 const LayoutOptionsContext = createContext<[number, ...number[]][]>([])
 
 export const LayoutOptionsProvider = LayoutOptionsContext.Provider
 
 // UI Components
-export const LayoutContainer = ({
+export function LayoutContainer ({
   attributes,
   children,
   element,
-}: RenderElementProps & { element: { type: 'layout' } }) => {
+}: RenderElementProps & { element: { type: 'layout' } }) {
   const { spacing } = useTheme()
   const focused = useFocused()
   const selected = useSelected()
@@ -114,7 +112,7 @@ export const LayoutContainer = ({
   )
 }
 
-export const LayoutArea = ({ attributes, children }: RenderElementProps) => {
+export function LayoutArea ({ attributes, children }: RenderElementProps) {
   const { colors, radii, spacing } = useTheme()
   return (
     <div
@@ -130,105 +128,6 @@ export const LayoutArea = ({ attributes, children }: RenderElementProps) => {
     </div>
   )
 }
-
-export const insertLayout = (editor: Editor, layout: [number, ...number[]]) => {
-  insertNodesButReplaceIfSelectionIsAtEmptyParagraphOrHeading(editor, [
-    {
-      type: 'layout',
-      layout,
-      children: [
-        { type: 'layout-area', children: [{ type: 'paragraph', children: [{ text: '' }] }] },
-      ],
-    },
-  ])
-  const layoutEntry = Editor.above(editor, { match: x => x.type === 'layout' })
-  if (layoutEntry) {
-    Transforms.select(editor, [...layoutEntry[1], 0])
-  }
-}
-
-// Plugin
-export function withLayouts (editor: Editor): Editor {
-  const { normalizeNode, deleteBackward } = editor
-  editor.deleteBackward = unit => {
-    if (
-      editor.selection &&
-      Range.isCollapsed(editor.selection) &&
-      // this is just an little optimisation
-      // we're only doing things if we're at the start of a layout area
-      // and the start of anything will always be offset 0
-      // so we'll bailout if we're not at offset 0
-      editor.selection.anchor.offset === 0
-    ) {
-      const [aboveNode, abovePath] = Editor.above(editor, {
-        match: node => node.type === 'layout-area',
-      }) || [editor, []]
-      if (
-        aboveNode.type === 'layout-area' &&
-        Point.equals(Editor.start(editor, abovePath), editor.selection.anchor)
-      ) {
-        return
-      }
-    }
-    deleteBackward(unit)
-  }
-  editor.normalizeNode = entry => {
-    const [node, path] = entry
-
-    if (Element.isElement(node) && node.type === 'layout') {
-      if (node.layout === undefined) {
-        Transforms.unwrapNodes(editor, { at: path })
-        return
-      }
-      if (node.children.length < node.layout.length) {
-        Transforms.insertNodes(
-          editor,
-          Array.from({
-            length: node.layout.length - node.children.length,
-          }).map(() => ({
-            type: 'layout-area',
-            children: [paragraphElement()],
-          })),
-          {
-            at: [...path, node.children.length],
-          }
-        )
-        return
-      }
-      if (node.children.length > node.layout.length) {
-        Array.from({
-          length: node.children.length - node.layout.length,
-        })
-          .map((_, i) => i)
-          .reverse()
-          .forEach(i => {
-            const layoutAreaToRemovePath = [...path, i + node.layout.length]
-            const child = node.children[i + node.layout.length] as Element
-            moveChildren(
-              editor,
-              layoutAreaToRemovePath,
-              [
-                ...path,
-                node.layout.length - 1,
-                (node.children[node.layout.length - 1] as Element).children.length,
-              ],
-              node => node.type !== 'paragraph' || Node.string(child) !== ''
-            )
-
-            Transforms.removeNodes(editor, {
-              at: layoutAreaToRemovePath,
-            })
-          })
-        return
-      }
-    }
-    normalizeNode(entry)
-  }
-  return editor
-}
-
-// Utils
-// ------------------------------
 
 function makeLayoutIcon (ratios: number[]) {
   const size = 16
@@ -255,7 +154,7 @@ function makeLayoutIcon (ratios: number[]) {
 
 const layoutsIcon = <ColumnsIcon size="small" />
 
-export const LayoutsButton = ({ layouts }: { layouts: DocumentFeatures['layouts'] }) => {
+export function LayoutsButton ({ layouts }: { layouts: DocumentFeatures['layouts'] }) {
   const {
     editor,
     layouts: { isSelected },
